处理id时间戳部分的类型为无符号方式记录32位，计算值作为64位，便于以后的兼容；

+++ contrib/base/nodeid.h
@@ -58,7 +58,7 @@ int nodeid_workerpart (int64 id);
 /**
  * get timestamp part.
  * */
-int nodeid_timestamp (int64 id);
+int64 nodeid_timestamp (int64 id);
 
 /**
  * get max nodeid.
